+++ shared/logutil/BUILD.bazel
@@ -5,9 +5,5 @@ go_library(
     srcs = ["logutil.go"],
     importpath = "github.com/prysmaticlabs/prysm/shared/logutil",
     visibility = ["//visibility:public"],
-    deps = [
-        "@com_github_sirupsen_logrus//:go_default_library",
-        "@org_golang_google_grpc//:go_default_library",
-        "@org_golang_google_grpc//metadata:go_default_library",
-    ],
+    deps = ["@com_github_sirupsen_logrus//:go_default_library"],
 )
